+++ server/persons.go
@@ -24,6 +24,7 @@ import (
 
 var personVersionConfig = entityVersionAPIConfig[vo.PersonVO, vo.PersonVersionVO]{
 	recordType: "person",
+	listPath:   "/persons",
 	get: func(c *gin.Context, id string) (*vo.PersonVO, error) {
 		return data.GetPerson(c.Request.Context(), id)
 	},
@@ -52,6 +53,12 @@ var personVersionConfig = entityVersionAPIConfig[vo.PersonVO, vo.PersonVersionVO
 	setCurrentVersion: func(c *gin.Context, id string, version int) (*vo.PersonVersionVO, error) {
 		return data.SetCurrentPersonVersion(c.Request.Context(), id, version)
 	},
+	softDelete: func(c *gin.Context, id string, deletedBy string) error {
+		return data.SoftDeletePerson(c.Request.Context(), id, deletedBy)
+	},
+	restore: func(c *gin.Context, id string) error {
+		return data.RestorePerson(c.Request.Context(), id)
+	},
 	versionState:  func(v *vo.PersonVersionVO) string { return string(v.State) },
 	versionNumber: func(v *vo.PersonVersionVO) int { return v.Version },
 	fields: map[string]entityFieldAccessor[vo.PersonVO]{
@@ -76,16 +83,18 @@ func setupPersonHandlers(g *gin.Engine, store persistence.CacheStore, ttls cache
 	g.GET("/persons/:id/versions/:version", getEntityVersion(personVersionConfig))
 
 	writeRoles := authz.RequireAnyRole(authzClient, constants.ServiceName, authz.RoleAdmin, authz.RoleEditor, authz.RoleSubmitter)
-	g.POST("/persons", writeRoles, createEntityVersion(personVersionConfig))
-	g.PATCH("/persons/:id", writeRoles, patchEntityVersion(personVersionConfig))
+	g.POST("/persons", writeRoles, createEntityVersion(personVersionConfig, store))
+	g.PATCH("/persons/:id", writeRoles, patchEntityVersion(personVersionConfig, store))
 	g.POST("/persons/:id/versions/:version/retract", writeRoles, retractEntityVersion(personVersionConfig))
 
 	reviewRoles := authz.RequireAnyRole(authzClient, constants.ServiceName, authz.RoleAdmin, authz.RoleEditor)
-	g.POST("/persons/:id/versions/:version/accept", reviewRoles, acceptEntityVersion(personVersionConfig))
+	g.POST("/persons/:id/versions/:version/accept", reviewRoles, acceptEntityVersion(personVersionConfig, store))
 	g.POST("/persons/:id/versions/:version/reject", reviewRoles, rejectEntityVersion(personVersionConfig))
 
 	rollbackRoles := authz.RequireAnyRole(authzClient, constants.ServiceName, authz.RoleAdmin)
 	g.POST("/persons/:id/versions/:version/current", rollbackRoles, setCurrentEntityVersion(personVersionConfig))
+	g.DELETE("/persons/:id", rollbackRoles, deleteEntity(personVersionConfig, store))
+	g.POST("/persons/:id/restore", rollbackRoles, restoreEntity(personVersionConfig, store))
 }
 
 // List persons.